rearrange the file structure
divide into models , services , controllers and separate the logic from the endpoint
