+++ Backend/employee-managment/employees.json
@@ -1,0 +1,82 @@
+[
+  {
+    "firstName": "Sayed",
+    "lastName": "Ahmed",
+    "employeeID": 1005,
+    "designation": "Back",
+    "knownLanguages": [
+      {
+        "scoreOutof100": 80,
+        "languageName": "Java"
+      },
+      {
+        "scoreOutof100": 75,
+        "languageName": "JavaScript"
+      },
+      {
+        "scoreOutof100": 85,
+        "languageName": "Python"
+      }
+    ]
+  },
+  {
+    "firstName": "John",
+    "lastName": "Doe",
+    "employeeID": 1001,
+    "designation": "Team Lead",
+    "knownLanguages": [
+      {
+        "scoreOutof100": 80,
+        "languageName": "Java"
+      },
+      {
+        "scoreOutof100": 75,
+        "languageName": "JavaScript"
+      },
+      {
+        "scoreOutof100": 85,
+        "languageName": "Python"
+      }
+    ]
+  },
+  {
+    "firstName": "John",
+    "lastName": "Doe",
+    "employeeID": 1001,
+    "designation": "Team Lead",
+    "knownLanguages": [
+      {
+        "scoreOutof100": 80,
+        "languageName": "Java"
+      },
+      {
+        "scoreOutof100": 75,
+        "languageName": "JavaScript"
+      },
+      {
+        "scoreOutof100": 85,
+        "languageName": "Python"
+      }
+    ]
+  },
+  {
+    "firstName": "John",
+    "lastName": "Doe",
+    "employeeID": 1001,
+    "designation": "Team Lead",
+    "knownLanguages": [
+      {
+        "scoreOutof100": 80,
+        "languageName": "Java"
+      },
+      {
+        "scoreOutof100": 75,
+        "languageName": "JavaScript"
+      },
+      {
+        "scoreOutof100": 85,
+        "languageName": "Python"
+      }
+    ]
+  }
+]
